backend/account/grpc_server/requirements.txt
grpcio==1.51.1
grpcio-tools==1.51.1
protobuf==4.21.12
psycopg2-binary==2.9.5
